Remove unsupported operations Move atomic to internal package
