+++ modules/fx/arrow-fx-reactor/src/main/kotlin/arrow/fx/reactor/MonoK.kt
@@ -4,7 +4,7 @@ import arrow.core.Either
 import arrow.core.Either.Left
 import arrow.core.Either.Right
 import arrow.core.NonFatal
-import arrow.core.extensions.AtomicBooleanW
+import arrow.core.internal.AtomicBooleanW
 import arrow.fx.OnCancel
 import arrow.fx.internal.Platform
 import arrow.fx.reactor.CoroutineContextReactorScheduler.asScheduler
