+++ modules/fx/arrow-fx-rx2/src/main/kotlin/arrow/fx/rx2/SingleK.kt
@@ -3,7 +3,7 @@ package arrow.fx.rx2
 import arrow.core.Either
 import arrow.core.Left
 import arrow.core.Right
-import arrow.core.extensions.AtomicRefW
+import arrow.core.internal.AtomicRefW
 import arrow.core.nonFatalOrThrow
 import arrow.fx.CancelToken
 import arrow.fx.internal.Platform
